Move testutils to allow to be reused outside of fiberapi
